+++ src/components/Forms/Controls/OptionCard/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/no-unknown-property */
 /** @jsxImportSource @emotion/react */
 
-import { RadioCard } from '@chakra-ui/react'
+import { RadioCard, HStack } from '@chakra-ui/react'
 
 import { OptionCardProps } from './types'
 import {
@@ -16,46 +16,58 @@ import {
 } from './styled'
 
 const OptionCard = ({
-  label,
-  caption,
-  icon,
-  variant = 'default',
-  disabled,
-  children,
-  value,
+  defaultValue,
+  items,
+  onValueChange,
 }: OptionCardProps) => (
-  <RadioCard.Item
-    css={optionCardContainerStyles}
-    value={value}
-    disabled={disabled}
-  >
-    <RadioCard.ItemHiddenInput />
-    <RadioCard.ItemControl css={optionCardControlStyles}>
-      <div css={optionCardHeaderStyles(variant)}>
-        <div css={optionCardIconContainerStyles}>{icon}</div>
-        <div style={{ width: '100%' }}>
-          <p
-            css={optionCardHeaderLabelStyles(variant, disabled)}
-            aria-label={label}
-            aria-disabled={disabled}
-          >
-            {label}
-          </p>
-          <p
-            css={optionCardHeaderCaptionStyles(variant, disabled)}
-            aria-label={caption}
-            aria-disabled={disabled}
-          >
-            {caption}
-          </p>
-        </div>
-      </div>
-      <RadioCard.ItemIndicator css={optionCardItemIndicatorStyles(variant)} />
-    </RadioCard.ItemControl>
-    {variant === 'expanded' ? (
-      <div css={optionCardExpandedContainerStyles(disabled)}>{children}</div>
-    ) : null}
-  </RadioCard.Item>
+  <RadioCard.Root defaultValue={defaultValue} onValueChange={onValueChange}>
+    <HStack alignItems='flex-start' flexWrap='wrap' gap='12px'>
+      {items.map((item) => (
+        <RadioCard.Item
+          css={optionCardContainerStyles}
+          key={`${item.label}-${item.value}`}
+          value={item.value}
+          disabled={item.disabled}
+        >
+          <RadioCard.ItemHiddenInput />
+          <RadioCard.ItemControl css={optionCardControlStyles}>
+            <div css={optionCardHeaderStyles(item.variant)}>
+              <div css={optionCardIconContainerStyles}>{item.icon}</div>
+              <div style={{ width: '100%' }}>
+                <p
+                  css={optionCardHeaderLabelStyles(item.variant, item.disabled)}
+                  aria-label={item.label}
+                  aria-disabled={item.disabled}
+                >
+                  {item.label}
+                </p>
+                <p
+                  css={optionCardHeaderCaptionStyles(
+                    item.variant,
+                    item.disabled,
+                  )}
+                  aria-label={item.caption}
+                  aria-disabled={item.disabled}
+                >
+                  {item.caption}
+                </p>
+              </div>
+            </div>
+            <RadioCard.ItemIndicator
+              css={optionCardItemIndicatorStyles(item.variant)}
+            />
+          </RadioCard.ItemControl>
+          {item.children && item.variant === 'expanded' ? (
+            <RadioCard.ItemAddon
+              css={optionCardExpandedContainerStyles(item.disabled)}
+            >
+              {item.children}
+            </RadioCard.ItemAddon>
+          ) : null}
+        </RadioCard.Item>
+      ))}
+    </HStack>
+  </RadioCard.Root>
 )
 
 export default OptionCard